Fix get_curve_{publickey,secretkey,serverkey}
For those, the size of the buffer passed to getsockopt is significant
and must be either 32 (to receive key as binary) or 41 (to receive
key as z85).

The patch merely sets the buffer size to 41 for those options, so that
the caller is returned the keys in z85 format.

Previously, calling those functions would result in an error with
errno set to EINVAL.

Closes #85