Quench warnings about safe-string, which returns const char*
In https://github.com/ocaml/ocaml/pull/1274 OCaml changed to return
`const char*` when safe-string is enabled as it is by default lately.
This means that our code throws warnings.

Some of the warnings can be avoided by replacing `memcpy` with
`caml_alloc_initialized_string` which is doing the right thing.

Other code needs to cast to access the destination of the pointer for
code that can't deal with `const char*`.